Not quite.  Right data type this time.

+++ Lab5/paging.c
@@ -10,9 +10,9 @@ int main(int argc, char * argv[]) {
       exit(-1);
    } 
    printf("Argv[1] %s\n",argv[1]);
-   unsigned long int address = atol(argv[1]); // convert string arg to int   
-   unsigned long int pageNumber = address / 4096; // 4kb page size
-   unsigned long int offset = address % 4096; // offset will be remainder
+   unsigned int address = atoi(argv[1]); // convert string arg to int   
+   unsigned int pageNumber = address / 4096; // 4kb page size
+   unsigned int offset = address % 4096; // offset will be remainder
    
    printf("The address %d contains:\npage number = %ld\noffset = %ld\n", address, pageNumber, offset);
